// src/presentation/app/crca/interfaces/crca.interface.ts
// Generated by https://quicktype.io

export interface Crca {
    id?:number;
    codigoCrca:           number;
    tipoCrca:             number;
    dniAportante:         string;
    nombreAportante:      string;
    tipoEspecie:          number;
    fechaRecepcion:       string | number | Date;
    valorNumero:          number;
    detalleAporte:        string;
    detalleAporteEspecie: string;
    estado:               number;
    estadoFormulario:     string;
}